requirements.txt
fastapi
uvicorn
pydantic
locust
requests
pyyaml
tabulate
prettytable
pandas
google-cloud-bigquery